added annotation @JsonIgnoreProperties to ignore properties on deserialization but not in serialization
it means now when system calls PUT and it has a unknown attribute, it will be ignored and exception appears. But when its a GET resource, this annotations allows attribute to be shown.
Name attribute e.g. It has to be shown on GET but has to be ignored on PUT resource